% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bland_altman_plot.R
\name{bland_altman_plot}
\alias{bland_altman_plot}
\title{Bland Altman Plot}
\usage{
bland_altman_plot(x, y)
}
\arguments{
\item{x}{a numeric vector (A)}

\item{y}{a numeric vector (B)}
}
\value{
a ggplot2 object
}
\description{
Bland Altman Plot
}
