% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{conf_int}
\alias{conf_int}
\title{Confidence Interval}
\usage{
conf_int(x, ci = 0.975)
}
\arguments{
\item{x}{a numeric vector}

\item{ci}{a numeric value (0 - 1) for the required confidence interval}
}
\value{
a numeric value for the lower and upper bounds of the confidence interval
}
\description{
Calculate a confidence interval for a vector of values
}
