% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int_tidy.R
\name{conf_int_tidy}
\alias{conf_int_tidy}
\title{Confidence Interval (Tidy)}
\usage{
conf_int_tidy(x, ci = 0.975)
}
\arguments{
\item{x}{a numeric vector}

\item{ci}{a numeric value (0 - 1) for the required confidence interval}
}
\value{
a character with the value for the lower and upper bounds of the confidence interval wrapped in parentheses for making tables
}
\description{
Calculate a confidence interval for a vector of values
}
