% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice_sorensen.R
\name{dice_sorensen}
\alias{dice_sorensen}
\title{Dice Sorensen Index}
\usage{
dice_sorensen(x, y)
}
\arguments{
\item{x}{a character vector}

\item{y}{a character vector}
}
\value{
a numeric value for the Dice-Sorensen Index
}
\description{
Calculate the Dice-Sorenson Index between two feature vectors
}
\references{
Zucknick, M., Richardson, S., Stronach, E.A., 2008. \emph{Comparing the characteristics of
gene expression profiles derived by univariate and multivariate classification
methods}. Statistical Applications in Genetics and Molecular Biology 7 (1):7

Loscalzo, S., Yu, L., Ding, C., 2009. \emph{Consensus group stable feature selection}. In:
Proceeding of the 15th ACM SIGKDD International Conference on Knowledge Discovery and Data Mining (KDD'09),pp.567-575.
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
