% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_accuracy.R
\name{forest_accuracy}
\alias{forest_accuracy}
\title{Forest Accuracy}
\usage{
forest_accuracy(model)
}
\arguments{
\item{model}{a randomForest model object}
}
\value{
a numeric value for the model accuracy (0 - 1)
}
\description{
Retrieve the forest accuracy (1 - OOB)
}
