% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_auc.R
\name{forest_auc}
\alias{forest_auc}
\title{Calculate the ROC-AUC from a randomForest model}
\usage{
forest_auc(model)
}
\arguments{
\item{model}{a randomForest classification model}
}
\value{
a numeric value for ROC-AUC
}
\description{
Calculte the training ROC-AUC for a randomForest model. The votes cast during out-of-bag (OOB) predictions are used for
the predicition of class. If classification is multinomial, then Hand and Till's (2001) method for multi-class AUC is used. If
classification is binary, then the standard ROC-AUC apprach is used.
}
