% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_feat_ranks.R
\name{forest_feat_ranks}
\alias{forest_feat_ranks}
\title{Random Forest feature rankings}
\usage{
forest_feat_ranks(model, meth = "gini")
}
\arguments{
\item{model}{a randomForest model object}

\item{meth}{string of either \code{gini} (for MeanDecreaseGini) or \code{perm} (for MeanDecreaseAcc)}
}
\value{
a \code{data.frame} of feature importance scores and rankings
}
\description{
Create rankings for random forest feature importance scores
}
