% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_kappa.R
\name{forest_kappa}
\alias{forest_kappa}
\title{Calculate Cohen's Kappa from a randomForest object}
\usage{
forest_kappa(model)
}
\arguments{
\item{model}{a randomForest classification model}
}
\value{
a numeric value for the overall inter-rate agreement (Kappa)
}
\description{
Calculte the training inter-rate agreement (Kappa) for a randomForest model. The votes cast during out-of-bag (OOB) predictions are used for
the predicition of class.
}
