% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_kappa.R
\name{get_test_kappa}
\alias{get_test_kappa}
\title{Calculate the test inter-rate agreement}
\usage{
get_test_kappa(train_model, test_data, test_cls)
}
\arguments{
\item{train_model}{a training model}

\item{test_data}{a code{data.frame} of data for test predictions}

\item{test_cls}{a vector of class labels for \code{test_data}}
}
\description{
Calculate Cohen's Kappa using a training model and independant test-data
}
