% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammings_distance.R
\name{hammings_distance}
\alias{hammings_distance}
\title{Relative Hamming Distance}
\usage{
hammings_distance(x, y, m)
}
\arguments{
\item{x}{a character vector}

\item{y}{a character vector o}

\item{m}{a numeric value for the total number of features in the dataset}
}
\value{
a numeric value for the Relative Hamming Distance
}
\description{
Calculate the Relative Hamming Distance between two feaure vectors
}
\references{
Dunne, K., Cunningham, P., Azuaje, F., 2002. \emph{Solutions to instability problems with
sequential wrapper-based approaches to feature selection}. Technical Report,Department of Computer Science, Trinity College, Dublin.
}
