% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccards_index.R
\name{jaccards_index}
\alias{jaccards_index}
\title{Jaccard's Similiarity}
\usage{
jaccards_index(x, y)
}
\arguments{
\item{x}{a character vector}

\item{y}{a character vector}
}
\value{
a numeric value for the Jaccard's Similiarity Coefficent
}
\description{
Calculate the Jaccard's Similiarity (or Taminoto Distance) between feature vectors
}
