% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ochiais_index.R
\name{ochiais_index}
\alias{ochiais_index}
\title{Ochiais Index}
\usage{
ochiais_index(x, y)
}
\arguments{
\item{x}{a character vector}

\item{y}{a character vector}
}
\value{
a numeric value for Ochiais Index
}
\description{
Calculate  Ochiais Index between feature vectors
}
