% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_overlap.R
\name{percentage_overlap}
\alias{percentage_overlap}
\title{Percentage overlap}
\usage{
percentage_overlap(x, y)
}
\arguments{
\item{x}{a character vector}

\item{y}{a character vector}
}
\value{
two numeric values. Value one is the percentage of \code{x} which overlaps with \code{y}.
Value two is the percentage of \code{y} which overlaps with \code{x}
}
\description{
Calculate the percentage of overlap between two feature vectors
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
