% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rf_confusion.R
\name{plot_rf_confusion}
\alias{plot_rf_confusion}
\title{Plot Confusion Matrix from Random Forest Classification}
\usage{
plot_rf_confusion(rf_model)
}
\arguments{
\item{rf_model}{a randomForest classification model}
}
\value{
a ggplot2 plot
}
\description{
Plot Confusion Matrix from Random Forest Classification
}
