% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity_to_mds.R
\name{proximity_to_mds}
\alias{proximity_to_mds}
\title{Multi Dimensional Scaling (MDS) of randomForest proximities}
\usage{
proximity_to_mds(x)
}
\arguments{
\item{x}{a `randomForest` object containing a valid proximity matrix}
}
\value{
a `data.frame` of `cmdscale (1 - proxmimity)` for Dimension 1 and 2
}
\description{
Multi Dimensional Scaling (MDS) of randomForest proximities
}
