% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_resamp.R
\name{strat_resamp}
\alias{strat_resamp}
\title{Stratified Resampling}
\usage{
strat_resamp(x, cls, p)
}
\arguments{
\item{x}{a \code{data.frame} of variables and observations}

\item{cls}{a vector of class information for stratifying. It is assumed that \code{cls} is balanaced}

\item{p}{a numeric value for the partitioning ratio (ie, 0.632)}
}
\value{
a list of four elements
  \itemize{
      \item{train_cls} \code{cls} vector for training set
      \item{train_x} training data
      \item{test_cls} \code{cls} vector for test set
      \item{test_x} test data
  }
}
\description{
Create a training and test set, stratfied by \code{class}
}
