% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varExp.R
\name{varExp}
\alias{varExp}
\title{Variance Explained}
\usage{
varExp(x)
}
\arguments{
\item{x}{a \code{prcomp} object}
}
\value{
a numeric vector of percentage variance explained for each principal componenet (PC)
}
\description{
Variance Explained
}
