% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_exp.R
\name{variance_exp}
\alias{variance_exp}
\title{Variance Explained}
\usage{
variance_exp(x)
}
\arguments{
\item{x}{a \code{prcomp} object}
}
\value{
a numeric vector of percentage variance explained for each principal componenet (PC)
}
\description{
Variance Explained
}
