% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity_to_mds.R
\name{proximity_to_mds}
\alias{proximity_to_mds}
\title{RF - MDS}
\usage{
proximity_to_mds(x)
}
\arguments{
\item{x}{a \code{randomForest} object containing a valid proximity matrix}
}
\value{
a \code{data.frame} of \code{cmdscale (1 - proxmimity)} for Dimension 1 and 2
}
\description{
Multi Dimensional Scaling (MDS) of randomForest proximities
}
