% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpt.R
\name{rpt}
\alias{rpt}
\title{Robustness Performance Trade Off}
\usage{
rpt(stability, performance, beta = 1)
}
\arguments{
\item{stability}{a numeric value for model stability}

\item{performance}{a numeric value for model performance}

\item{beta}{a positive integer.  Default is 1, which treats stability and performance equally.}
}
\value{
a numeric value for RPT between 0 and 1
}
\description{
Calculates the Robustness Performance Trade-off of a model
}
\author{
Tom Wilson
}
